Updated release path setting
Prior, the release path assumed a directory and appended
`start_lexical.sh` to it. This meant that the name of the executable was
hardcoded in the extension, and that made playing around with packaging
tougher. It would also mean that getting lexical to run on windows would
be harder.

This PR maintains backwards compatibility with the old pattern where you
can specify a directory, but also adds the ability to point to an
executable file.